requirements.txt
pywebview==4.0
tk
keyboard
screeninfo